SampleApp Version Final

+++ app/helpers/sessions_helper.rb
@@ -4,6 +4,7 @@ module SessionsHelper
   end
 
   def remember(user)
+    return unless user
     user.remember
     cookies.permanent.encrypted[:user_id] = user.id
     cookies.permanent[:remember_token] = user.remember_token
@@ -26,12 +27,14 @@ module SessionsHelper
   end
 
   def forget(user)
+    return unless user
     user.forget
     cookies.delete(:user_id)
     cookies.delete(:remember_token)
   end
 
   def log_out
+    return unless current_user
     forget(current_user)
     session.delete(:user_id)
     @current_user = nil
